+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - [fix:fix nacos and consul registration.](https://github.com/Tencent/spring-cloud-tencent/pull/922)
 - [Documentation content changes: add circuitbreaker readme.](https://github.com/Tencent/spring-cloud-tencent/pull/931)
 - [fix: fix PolarisRouterServiceInstanceListSupplier npe with reactive feign.](https://github.com/Tencent/spring-cloud-tencent/pull/927)
+- [fix: adapt for nacos instance.](https://github.com/Tencent/spring-cloud-tencent/pull/936)